Keep Jedis target creation in singletons
