--- instrumentation/jedis/jedis-1.4/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jedis/v1_4/JedisServerTargets.java
@@ -1,30 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.jedis.v1_4;
-
-import io.opentelemetry.instrumentation.api.incubator.semconv.db.internal.RedisServerTarget;
-import java.util.ArrayList;
-import java.util.List;
-import javax.annotation.Nullable;
-import redis.clients.jedis.JedisShardInfo;
-
-public class JedisServerTargets {
-
-  @Nullable
-  public static RedisServerTarget ofShards(@Nullable List<JedisShardInfo> shards) {
-    if (shards == null || shards.isEmpty()) {
-      return null;
-    }
-    List<String> endpoints = new ArrayList<>(shards.size());
-    for (JedisShardInfo shard : shards) {
-      endpoints.add(
-          shard == null ? null : RedisServerTarget.endpoint(shard.getHost(), shard.getPort()));
-    }
-    return RedisServerTarget.ofEndpoints(endpoints);
-  }
-
-  private JedisServerTargets() {}
-}
