+++ instrumentation/jedis/jedis-1.4/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jedis/v1_4/ShardedJedisInstrumentation.java
@@ -37,7 +37,7 @@ class ShardedJedisInstrumentation implements TypeInstrumentation {
 
     @Advice.OnMethodEnter(suppress = Throwable.class, inline = false)
     public static Scope onEnter(@Advice.Argument(0) @Nullable List<JedisShardInfo> shards) {
-      return JedisSingletons.openConfiguredTargetScope(JedisServerTargets.ofShards(shards));
+      return JedisSingletons.openConfiguredTargetScope(JedisSingletons.createServerTarget(shards));
     }
 
     @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class, inline = false)
